feat: created folder for pages
