+++ src/pages/Home.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Header from "../components/Header";
+
+function Home() {
+    return (
+        <>
+           <p>Home</p> 
+        </>
+    )
+}
+
+export default Home
