+++ src/pages/SignIn.jsx
@@ -1,0 +1,9 @@
+function SignIn() {
+  return (
+      <>
+        <p>Sign in</p>
+      </>
+    );
+}
+
+export default SignIn;
